Add files via upload
In these programs, we  tried to do some basic calculations by using c language. 
Some of these programs are used to convert the value of temperature to \ from Fahrenheit to\from Celsius and
Some of these codes are used to find the area of triangle and volume of sphere.
In our new videos,
We will try to forward our step in next level.